pyproject.toml
[project]
name = "corridorkey"
version = "1.0.0"
description = "Neural network green screen keying for professional VFX pipelines"
readme = "README.md"
requires-python = ">=3.10, <=3.14"
license = { text = "CC-BY-NC-SA-4.0" }
authors = [{ name = "Corridor Digital" }]

dependencies = [
    # Core inference
    "torch==2.10.0",
    "torchvision==0.25.0",
    "timm==1.0.24",
    "numpy",
    "opencv-python",
    "tqdm",
    "setuptools",
    # Triton for torch.compile (Windows needs a dedicated package)
    "triton-windows==3.6.0.post25 ; sys_platform == 'win32'",

    # GVM alpha hint generator
    "diffusers",
    "transformers",
    "accelerate",
    "peft",
    "av",
    "Pillow",
    "PIMS",
    "easydict",
    "imageio",
    "matplotlib",

    # BiRefNet alpha hint generator (lightweight, ~4GB VRAM, MIT license)
    "kornia",

    # VideoMaMa alpha hint generator
    "einops",

    # CLI tools (huggingface-hub is also a transitive dep, but must be direct
    # so that uv installs the "hf" console-script entry point)
    "huggingface-hub",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

# TensorRT (optional): install torch-tensorrt manually to match your torch version.
# See https://pytorch.org/TensorRT/getting_started/installation.html
# Usage: uv pip install torch-tensorrt && corridorkey --tensorrt

[project.scripts]
corridorkey = "corridorkey_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["CorridorKeyModule", "gvm_core", "birefnet_core", "VideoMaMaInferenceModule"]

[tool.ruff]
target-version = "py311"
line-length = 120
extend-exclude = ["gvm_core/", "VideoMaMaInferenceModule/"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations
    "RUF",    # ruff-specific rules
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PIE",    # flake8-pie
    "RET",    # flake8-return
    "C4",     # flake8-comprehensions
    "PT",     # flake8-pytest-style
    "PYI",    # flake8-pyi
    "T20",    # flake8-print
]
ignore = [
    "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed
    "RET504",  # unnecessary assignment before return
    "SIM108",  # use ternary operator (often less readable)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "T20"]
"corridorkey_cli.py" = ["T20"]  # CLI prints are expected
"benchmark_4k_vram.py" = ["T20", "ANN"]
"test_vram.py" = ["T20", "ANN"]
"download_test_scene.py" = ["T20", "ANN"]
"tears_of_steel_test/*" = ["T20", "ANN"]

[tool.ruff.format]
# Third-party code excluded via top-level exclude

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "gpu: requires CUDA GPU (skipped in CI)",
    "slow: long-running test",
    "mlx: requires Apple Silicon with MLX installed",
]
addopts = "--tb=short"

[tool.coverage.run]
branch = true
# "." covers root-level modules (clip_manager.py, corridorkey_cli.py once PR #8 merges).
# Third-party and non-library code is excluded via omit.
source = ["CorridorKeyModule", "."]
omit = [
    "gvm_core/*",
    "VideoMaMaInferenceModule/*",
    "tests/*",
    "test_vram.py",  # manual GPU smoke test, not part of the pytest suite
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"  # cu128 required for RTX 5000 series
explicit = true

[tool.uv.sources]
torch = { index = "pytorch", marker = "sys_platform != 'darwin'" }
torchvision = { index = "pytorch", marker = "sys_platform != 'darwin'" }
# Hiera FlashAttention fix: upstream timm expects 5D attention tensors,
# this fork produces contiguous 4D tensors for PyTorch's fused SDPA kernel
timm = { git = "https://github.com/Raiden129/pytorch-image-models-fix", branch = "fix/hiera-flash-attention-global-4d" }
